+++ lib/karafka/app.rb
@@ -9,9 +9,10 @@ module Karafka
       # Sets up all the internal components and bootstrap whole app
       # We need to know details about consumers in order to setup components,
       # that's why we don't setup them after std setup is done
-      # @raise [Karafka::Errors::InvalidConfiguration] raised when configuration
+      # @raise [Karafka::Errors::InvalidConfigurationError] raised when configuration
       #   doesn't match with ConfigurationSchema
       def boot!
+        initialize!
         Setup::Config.validate!
         Setup::Config.setup_components
         events.publish('after_init', config: Karafka::App.config)