+++ lib/karafka/base_responder.rb
@@ -141,7 +141,7 @@ module Karafka
 
       return if result.success?
 
-      raise Karafka::Errors::InvalidResponderUsage, result.errors
+      raise Karafka::Errors::InvalidResponderUsageError, result.errors
     end
 
     # Checks if we met all the options requirements before sending them to the producer.
@@ -152,7 +152,7 @@ module Karafka
         messages_set.each do |message_data|
           result = self.class.options_schema.call(message_data.last)
           next if result.success?
-          raise Karafka::Errors::InvalidResponderMessageOptions, result.errors
+          raise Karafka::Errors::InvalidResponderMessageOptionsError, result.errors
         end
       end
     end