+++ lib/karafka/connection/api_adapter.rb
@@ -87,7 +87,11 @@ module Karafka
           [
             Karafka::App.config.topic_mapper.outgoing(topic),
             partition,
-            { timeout: consumer_group.pause_timeout }
+            {
+              timeout: consumer_group.pause_timeout,
+              max_timeout: consumer_group.pause_max_timeout,
+              exponential_backoff: consumer_group.pause_exponential_backoff
+            }
           ]
         end
 
@@ -121,7 +125,7 @@ module Karafka
             next unless AttributesMap.api_adapter[namespace_key].include?(setting_name)
             # Ignore settings that are already initialized
             # In case they are in preexisting settings fetched differently
-            next if preexisting_settings.keys.include?(setting_name)
+            next if preexisting_settings.key?(setting_name)
             # Fetch all the settings from a given layer object. Objects can handle the fallback
             # to the kafka settings, so
             preexisting_settings[setting_name] = route_layer.send(setting_name)