+++ spec/lib/karafka/base_consumer_spec.rb
@@ -105,4 +105,38 @@ RSpec.describe Karafka::BaseConsumer do
       base_consumer.send(:mark_as_consumed, params)
     end
   end
+
+  describe '#mark_as_consumed!' do
+    let(:client) { instance_double(Karafka::Connection::Client) }
+    let(:params) { instance_double(Karafka::Params::Params) }
+
+    before { Karafka::Persistence::Client.write(client) }
+
+    it 'expect to proxy pass to client' do
+      expect(client).to receive(:mark_as_consumed!).with(params)
+      base_consumer.send(:mark_as_consumed!, params)
+    end
+  end
+
+  describe 'trigger_heartbeat' do
+    let(:client) { instance_double(Karafka::Connection::Client) }
+
+    before { Karafka::Persistence::Client.write(client) }
+
+    it 'expect to proxy pass to client' do
+      expect(client).to receive(:trigger_heartbeat)
+      base_consumer.send(:trigger_heartbeat)
+    end
+  end
+
+  describe 'trigger_heartbeat!' do
+    let(:client) { instance_double(Karafka::Connection::Client) }
+
+    before { Karafka::Persistence::Client.write(client) }
+
+    it 'expect to proxy pass to client' do
+      expect(client).to receive(:trigger_heartbeat!)
+      base_consumer.send(:trigger_heartbeat!)
+    end
+  end
 end