+++ spec/lib/karafka/connection/batch_delegator_spec.rb
@@ -1,13 +1,21 @@
 # frozen_string_literal: true
 
-RSpec.describe Karafka::Connection::Delegator do
+RSpec.describe Karafka::Connection::BatchDelegator do
   subject(:delegator) { described_class }
 
   let(:group_id) { consumer_group.id }
   let(:topic_id) { consumer_group.topics[0].name }
   let(:consumer_instance) { consumer_group.topics[0].consumer.new }
-  let(:messages_batch) { [raw_message1, raw_message2] }
+  let(:kafka_messages) { [raw_message1, raw_message2] }
   let(:raw_message_value) { rand }
+  let(:kafka_batch) do
+    instance_double(
+      Kafka::FetchedBatch,
+      messages: kafka_messages,
+      topic: topic_id,
+      partition: 0
+    )
+  end
   let(:raw_message1) do
     Kafka::FetchedMessage.new(
       message: OpenStruct.new(
@@ -61,23 +69,15 @@ RSpec.describe Karafka::Connection::Delegator do
     end
 
     it 'expect to run without errors' do
-      expect { delegator.call(group_id, messages_batch) }.not_to raise_error
+      expect { delegator.call(group_id, kafka_batch) }.not_to raise_error
     end
   end
 
   context 'when batch_consuming false' do
     before do
-      expect(consumer_instance)
-        .to receive(:params_batch=)
-        .with([raw_message1])
-
-      expect(consumer_instance)
-        .to receive(:params_batch=)
-        .with([raw_message2])
-
-      expect(consumer_instance)
-        .to receive(:call)
-        .twice
+      allow(consumer_instance).to receive(:params_batch=).with([raw_message1])
+      allow(consumer_instance).to receive(:params_batch=).with([raw_message2])
+      allow(consumer_instance).to receive(:call).twice
     end
 
     let(:consumer_group) do
@@ -93,7 +93,7 @@ RSpec.describe Karafka::Connection::Delegator do
     end
 
     it 'expect to run without errors' do
-      expect { delegator.call(group_id, messages_batch) }.not_to raise_error
+      expect { delegator.call(group_id, kafka_batch) }.not_to raise_error
     end
   end
 end