+++ spec/lib/karafka/connection/client_spec.rb
@@ -51,17 +51,37 @@ RSpec.describe Karafka::Connection::Client do
   end
 
   describe '#pause' do
+    let(:pause_timeout) { rand }
+    let(:pause_max_timeout) { rand }
+    let(:pause_exponential_backoff) { false }
+    let(:pause_args) do
+      [
+        topic,
+        partition,
+        timeout: pause_timeout,
+        max_timeout: pause_max_timeout,
+        exponential_backoff: pause_exponential_backoff
+      ]
+    end
+
     before do
       client.instance_variable_set(:'@kafka_consumer', kafka_consumer)
-      allow(consumer_group).to receive(:pause_timeout).and_return(pause_timeout)
+      allow(consumer_group)
+        .to receive(:pause_timeout)
+        .and_return(pause_timeout)
+      allow(consumer_group)
+        .to receive(:pause_max_timeout)
+        .and_return(pause_max_timeout)
+      allow(consumer_group)
+        .to receive(:pause_exponential_backoff)
+        .and_return(pause_exponential_backoff)
     end
 
     context 'when pause_timeout is set to nil' do
-      let(:pause_timeout) { nil }
       let(:error) { Karafka::Errors::InvalidPauseTimeout }
 
       it 'expect to raise an exception' do
-        expect(kafka_consumer).to receive(:pause).with(topic, partition, timeout: pause_timeout)
+        expect(kafka_consumer).to receive(:pause).with(*pause_args)
         expect(client.pause(topic, partition)).to eq true
       end
     end
@@ -71,7 +91,7 @@ RSpec.describe Karafka::Connection::Client do
       let(:error) { Karafka::Errors::InvalidPauseTimeout }
 
       it 'expect to raise an exception' do
-        expect(kafka_consumer).to receive(:pause).with(topic, partition, timeout: pause_timeout)
+        expect(kafka_consumer).to receive(:pause).with(*pause_args)
         expect(client.pause(topic, partition)).to eq true
       end
     end
@@ -80,7 +100,7 @@ RSpec.describe Karafka::Connection::Client do
       let(:pause_timeout) { rand(1..100) }
 
       it 'expect to pause consumer_group' do
-        expect(kafka_consumer).to receive(:pause).with(topic, partition, timeout: pause_timeout)
+        expect(kafka_consumer).to receive(:pause).with(*pause_args)
         expect(client.pause(topic, partition)).to eq true
       end
     end
@@ -95,13 +115,22 @@ RSpec.describe Karafka::Connection::Client do
           end
         end
       end
+      let(:pause_args) do
+        [
+          r_topic,
+          partition,
+          timeout: pause_timeout,
+          max_timeout: pause_max_timeout,
+          exponential_backoff: pause_exponential_backoff
+        ]
+      end
 
       before do
         allow(Karafka::App.config).to receive(:topic_mapper).and_return(custom_mapper)
       end
 
       it 'expect to pause consumer_group for a remapped topic' do
-        expect(kafka_consumer).to receive(:pause).with(r_topic, partition, timeout: pause_timeout)
+        expect(kafka_consumer).to receive(:pause).with(*pause_args)
         expect(client.pause(topic, partition)).to eq true
       end
     end
@@ -112,10 +141,40 @@ RSpec.describe Karafka::Connection::Client do
 
     before { client.instance_variable_set(:'@kafka_consumer', kafka_consumer) }
 
+    it 'expect to forward to mark_message_as_processed and not to commit offsets' do
+      expect(kafka_consumer).to receive(:mark_message_as_processed).with(params)
+      expect(kafka_consumer).not_to receive(:commit_offsets)
+      client.mark_as_consumed(params)
+    end
+  end
+
+  describe '#mark_as_consumed!' do
+    let(:params) { instance_double(Karafka::Params::Params) }
+
+    before { client.instance_variable_set(:'@kafka_consumer', kafka_consumer) }
+
     it 'expect to forward to mark_message_as_processed and commit offsets' do
       expect(kafka_consumer).to receive(:mark_message_as_processed).with(params)
       expect(kafka_consumer).to receive(:commit_offsets)
-      client.mark_as_consumed(params)
+      client.mark_as_consumed!(params)
+    end
+  end
+
+  describe '#trigger_heartbeat' do
+    before { client.instance_variable_set(:'@kafka_consumer', kafka_consumer) }
+
+    it 'expect to use the consumers non blocking trigger_heartbeat method' do
+      expect(kafka_consumer).to receive(:trigger_heartbeat)
+      client.trigger_heartbeat
+    end
+  end
+
+  describe '#trigger_heartbeat!' do
+    before { client.instance_variable_set(:'@kafka_consumer', kafka_consumer) }
+
+    it 'expect to use the consumers blocking trigger_heartbeat! method' do
+      expect(kafka_consumer).to receive(:trigger_heartbeat!)
+      client.trigger_heartbeat!
     end
   end
 
@@ -126,25 +185,21 @@ RSpec.describe Karafka::Connection::Client do
 
     context 'when everything works smooth' do
       context 'with single message consumption mode' do
-        let(:messages) { [incoming_message] }
-
         it 'expect to use kafka_consumer to get each message and yield as an array of messages' do
           expect(kafka_consumer).to receive(:each_message).and_yield(incoming_message)
-          expect { |block| client.fetch_loop(&block) }.to yield_with_args(messages)
+          expect { |block| client.fetch_loop(&block) }
+            .to yield_with_args(incoming_message, :message)
         end
       end
 
       context 'with message batch consumption mode' do
         let(:batch_fetching) { true }
         let(:incoming_batch) { instance_double(Kafka::FetchedBatch) }
-        let(:incoming_messages) { [incoming_message, incoming_message] }
 
         it 'expect to use kafka_consumer to get messages and yield all of them' do
           expect(kafka_consumer).to receive(:each_batch).and_yield(incoming_batch)
-          expect(incoming_batch).to receive(:messages).and_return(incoming_messages)
-
           expect { |block| client.fetch_loop(&block) }
-            .to yield_successive_args(incoming_messages)
+            .to yield_successive_args([incoming_batch, :batch])
         end
       end
     end
@@ -160,7 +215,7 @@ RSpec.describe Karafka::Connection::Client do
 
       before do
         count = 0
-        allow(kafka_consumer).to receive(:each_message).exactly(2).times do
+        allow(kafka_consumer).to receive(:each_message).twice do
           count += 1
           count == 1 ? raise(error) : true
         end
@@ -182,7 +237,7 @@ RSpec.describe Karafka::Connection::Client do
 
       before do
         count = 0
-        allow(kafka_consumer).to receive(:each_message).exactly(2).times do
+        allow(kafka_consumer).to receive(:each_message).twice do
           count += 1
           count == 1 ? raise(error) : true
         end
@@ -210,7 +265,7 @@ RSpec.describe Karafka::Connection::Client do
     end
 
     context 'when kafka_consumer is not yet built' do
-      let(:kafka) { instance_double(Kafka::Client) }
+      let(:kafka_client) { instance_double(Kafka::Client) }
       let(:kafka_consumer) { instance_double(Kafka::Consumer) }
       let(:subscribe_params) do
         [
@@ -220,23 +275,10 @@ RSpec.describe Karafka::Connection::Client do
         ]
       end
 
-      before do
-        allow(Kafka)
-          .to receive(:new)
-          .with(
-            ::Karafka::App.config.kafka.seed_brokers,
-            logger: ::Karafka.logger,
-            client_id: ::Karafka::App.config.client_id,
-            socket_timeout: 30,
-            connect_timeout: 10,
-            sasl_plain_authzid: '',
-            ssl_ca_certs_from_system: false
-          )
-          .and_return(kafka)
-      end
+      before { allow(Karafka::Connection::Builder).to receive(:call).and_return(kafka_client) }
 
       it 'expect to build it and subscribe' do
-        expect(kafka).to receive(:consumer).and_return(kafka_consumer)
+        expect(kafka_client).to receive(:consumer).and_return(kafka_consumer)
         expect(kafka_consumer).to receive(:subscribe).with(*subscribe_params)
         expect(client.send(:kafka_consumer)).to eq kafka_consumer
       end