+++ spec/lib/karafka/routing/builder_spec.rb
@@ -10,6 +10,7 @@ RSpec.describe Karafka::Routing::Builder do
   ].freeze
 
   before { builder.clear }
+
   after { builder.clear }
 
   describe '#draw' do
@@ -152,7 +153,21 @@ RSpec.describe Karafka::Routing::Builder do
         end
       end
 
-      it { expect { invalid_route }.to raise_error(Karafka::Errors::InvalidConfiguration) }
+      it { expect { invalid_route }.to raise_error(Karafka::Errors::InvalidConfigurationError) }
+    end
+
+    context 'when we define multiple consumer groups and one is without topics' do
+      subject(:drawing) do
+        described_class.instance.draw do
+          consumer_group :group_name1 do
+            topic(:topic_name1) { consumer Class.new(Karafka::BaseConsumer) }
+          end
+
+          consumer_group(:group_name2) {}
+        end
+      end
+
+      it { expect { drawing }.to raise_error(Karafka::Errors::InvalidConfigurationError) }
     end
   end